+ moved ErrorSnackbar.tsx + moved Login.tsx
